+++ src/components/Badge/Badge.js
@@ -1,22 +1,65 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import classNames from 'classnames';
+import styled, { css } from 'styled-components';
+import { pulseInfinite } from '../../animations';
 
-import './Badge.scss';
+const cssPrimary = css`
+  background-color: ${p => p.theme.primary};
+`;
 
-const Badge = ({ appearance, children, animated }) => {
-  const classes = classNames({
-    'ft--badge': true,
-    [`ft--badge--${appearance}`]: true,
-    'ft--badge--animated': animated
-  });
+const cssSecondary = css`
+  background-color: ${p => p.theme.grey03};
+`;
 
-  return (
-    <span className={classes}>
-      {children}
-    </span>
-  );
-};
+const cssSuccess = css`
+  background-color: ${p => p.theme.success};
+`;
+
+const cssDanger = css`
+  background-color: ${p => p.theme.danger};
+`;
+
+const cssWarning = css`
+  background-color: ${p => p.theme.warning};
+`;
+
+const cssLight = css`
+  background-color: ${p => p.theme.backgroundSecondary};
+  color: ${p => p.theme.grey01};
+`;
+
+const cssDark = css`
+  background-color: ${p => p.theme.grey01};
+`;
+
+const cssAnimated = css`
+  animation: ${pulseInfinite} 3s infinite ease;
+`;
+
+const BadgeWrapper = styled('span')`
+  display: inline-block;
+  padding: 0.2em 0.8em;
+  font-size: 80%;
+  border-radius: 1em;
+  text-align: center;
+  color: ${p => p.theme.white};
+  background-color: ${p => p.theme.primary};
+  line-height: 1;
+  ${p => p.appearance === 'primary' && cssPrimary};
+  ${p => p.appearance === 'secondary' && cssSecondary};
+  ${p => p.appearance === 'success' && cssSuccess};
+  ${p => p.appearance === 'danger' && cssDanger};
+  ${p => p.appearance === 'warning' && cssWarning};
+  ${p => p.appearance === 'light' && cssLight};
+  ${p => p.appearance === 'dark' && cssDark};
+  ${p => p.animated && cssAnimated};
+`;
+
+const Badge = ({ appearance, children, animated }) => (
+  <BadgeWrapper appearance={appearance} animated={animated}>
+    {children}
+  </BadgeWrapper>
+);
 
 Badge.defaultProps = {
   appearance: 'primary',